+++ src/Oklab.h
@@ -4,10 +4,10 @@
 #include "MatrixOperations.h"
 
 using RGB = std::array<int, 3>;
-using GsRGB = std::array<double, 3>;
-using LsRGB = std::array<double, 3>;
-using Oklab = std::array<double, 3>;
-using Oklch = std::array<double, 3>;
+using GsRGB = std::array<float, 3>;
+using LsRGB = std::array<float, 3>;
+using Oklab = std::array<float, 3>;
+using Oklch = std::array<float, 3>;
 
 RGB scaleTo255(const GsRGB &color);
 GsRGB scaleTo01(const RGB &color);
@@ -19,5 +19,5 @@ Oklab linearSrgbToOklab(const LsRGB &color);
 LsRGB oklabToLinearSrgb(const Oklab &color);
 Oklab oklchToOklab(const Oklch &color);
 Oklch oklabToOklch(const Oklab &color);
-double delta(const Oklab &c1, const Oklab &c2);
-RGB interpolateColor(const RGB &c1, const RGB &c2, double t);
+float delta(const Oklab &c1, const Oklab &c2);
+RGB interpolateColor(const RGB &c1, const RGB &c2, float t);
